improve test speed
